update: separate styled components from Admin / AdminTabs
